Another approach to mitigate #17480 on Windows
Previously we installed `process` in the user package database. However,
this is quite fragile and in fact breaks on Hadrian (presumably because
it is careful about hiding the user package database during the
bootstrap build). Instead we now treat `process` as a proper bootstrap
package in both the Hadrian and make build systems.

Fixes #17480.