simplCore: Ensure that rule templates contain no ticks
This fixes #17619, where a tick snuck in to the template of a rule,
resulting in a panic during rule matching. The situation is summarized
in `Note [Rule templates are devoid of ticks]`. In short, we ensure that
the template remains devoid of ticks by stripping ticks after
substitution. This feels a bit hacky but I can't think of a better
option at the moment.

We also now assert this invariant in Core Lint.